requirements.txt
any-agent==0.21.1
langchain==0.3.26
openai==1.90.0
jinja2==3.1.6
pydantic==2.11.7
dateparser==1.2.1
PyYAML==6.0.2
langdetect==1.0.9
google-auth==2.40.3
google-auth-oauthlib==1.2.2
google-api-python-client==2.173.0
ics==0.7.2
python-dotenv==1.1.0
pytest==8.4.1
setuptools==80.9.0
fastapi[standard]==0.115.13
apscheduler==3.11.0
langchain-community==0.3.26
langchain-openai==0.3.24
beautifulsoup4==4.13.4
pytesseract>=0.3.10
Pillow>=10.0.0
notion-client==2.4.0
any-agent[langchain]